address some self-review comments
